Setting file migration

+++ migrations/20190830225858-add_isStared_to_menus.js
@@ -13,7 +13,7 @@ module.exports = {
       queryInterface.addColumn(
         'menus',
         'isStared',
-        Sequelize.TINYINT
+        Sequelize.INTEGER(1)
       )
     ]);
   },
